requirements.txt
boto
bz2file
Cython
gensim
mecab-python3
numpy
requests
scipy
six
smart-open
